Removed MaterialInspector, MaterialView moved to PropInspector
