+++ Editor/App.cpp
@@ -15,7 +15,7 @@
 #include "Gizmo.h"
 #include "Global.h"
 #include "Grid.h"
-#include "MaterialInspector.h"
+#
 #include "Mod.h"
 #include "Node.h"
 #include "OutlinerWindow.h"
@@ -622,10 +622,6 @@ namespace ToolKit
         inspector->m_name        = g_propInspector;
         m_windows.push_back(inspector);
 
-        MaterialInspector* matInspect = new MaterialInspector();
-        matInspect->m_name            = g_matInspector;
-        m_windows.push_back(matInspect);
-
         PluginWindow* plugWindow = new PluginWindow();
         m_windows.push_back(plugWindow);
 
@@ -677,17 +673,11 @@ namespace ToolKit
           case Window::Type::Inspector:
             wnd = new PropInspector(wndNode);
             break;
-          case Window::Type::MaterialInspector:
-            wnd = new MaterialInspector(wndNode);
-            break;
           case Window::Type::PluginWindow:
             wnd = new PluginWindow(wndNode);
             break;
           case Window::Type::Viewport2d:
             wnd = new EditorViewport2d(wndNode);
-            break;
-          default:
-            assert(false);
             break;
           }
 
@@ -1148,11 +1138,6 @@ namespace ToolKit
     PropInspector* App::GetPropInspector()
     {
       return GetWindow<PropInspector>(g_propInspector);
-    }
-
-    MaterialInspector* App::GetMaterialInspector()
-    {
-      return GetWindow<MaterialInspector>(g_matInspector);
     }
 
     void App::RenderSelected(EditorViewport* viewport,
