+++ Editor/Editor.vcxproj
@@ -117,7 +117,6 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Use</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Use</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="MaterialInspector.cpp" />
     <ClCompile Include="MaterialView.cpp" />
     <ClCompile Include="OutlinerWindow.cpp" />
     <ClCompile Include="Pass.cpp" />
@@ -181,7 +180,6 @@
     <ClInclude Include="Global.h" />
     <ClInclude Include="Grid.h" />
     <ClInclude Include="ImGui\imgui_impl_opengl3_loader.h" />
-    <ClInclude Include="MaterialInspector.h" />
     <ClInclude Include="MaterialView.h" />
     <ClInclude Include="Pass.h" />
     <ClInclude Include="PluginWindow.h" />
