+++ Editor/Editor.vcxproj.filters
@@ -111,9 +111,6 @@
     <ClCompile Include="Action.cpp">
       <Filter>Source</Filter>
     </ClCompile>
-    <ClCompile Include="MaterialInspector.cpp">
-      <Filter>UI</Filter>
-    </ClCompile>
     <ClCompile Include="AnchorMod.cpp">
       <Filter>Mods</Filter>
     </ClCompile>
@@ -238,9 +235,6 @@
     </ClInclude>
     <ClInclude Include="Action.h">
       <Filter>Source</Filter>
-    </ClInclude>
-    <ClInclude Include="MaterialInspector.h">
-      <Filter>UI</Filter>
     </ClInclude>
     <ClInclude Include="AnchorMod.h">
       <Filter>Mods</Filter>
