--- Editor/MaterialInspector.h
@@ -1,45 +1,0 @@
-#pragma once
-
-#include "PropInspector.h"
-
-namespace ToolKit
-{
-  namespace Editor
-  {
-
-    class MaterialView : public View
-    {
-     public:
-      MaterialView() : View("Material Inspector")
-      {
-        m_viewID = 3;
-      }
-      virtual ~MaterialView()
-      {
-      }
-      void Show() override;
-
-     public:
-      MaterialPtr m_material;
-    };
-
-    class MaterialInspector : public Window
-    {
-     public:
-      explicit MaterialInspector(XmlNode* node);
-      MaterialInspector();
-      virtual ~MaterialInspector();
-
-      void Show() override;
-      Type GetType() const override;
-      void DispatchSignals() const override;
-
-     public:
-      MaterialPtr m_material;
-
-     private:
-      MaterialView* m_view;
-    };
-
-  } // namespace Editor
-} // namespace ToolKit
