+++ Editor/MaterialView.h
@@ -5,7 +5,16 @@ namespace ToolKit
 {
   namespace Editor
   {
+    class MaterialView : public View
+    {
+     public:
+      MaterialView();
 
+      void Show() override;
+
+     public:
+      MaterialPtr m_material;
+    };
 
   }
 }
